requirements.txt
google-generativeai
python-dotenv
langchain
langchain_community
PyPDF2
#chromadb
faiss-cpu
langchain-google-genai
Markdown
streamlit
streamlit-authenticator
